feat(therapist dashboard): creat a sidebar component with test and use context to manage state
