+++ src/components/Sidebar/__test__/Sidebar.test.js
@@ -1,0 +1,7 @@
+import renderer from "react-test-renderer";
+import Sidebar from "../Sidebar";
+
+it("renders correctly", () => {
+    const tree = renderer.create(<Sidebar />).toJSON();
+    expect(tree).toMatchSnapshot();
+});
